+++ testfiles/calls.py
@@ -1,4 +1,6 @@
 print('Hello world')
 
-foo()
-foo()(bar)
+foo(bar, a=None, x=1)
+foo()(bar, bang, x=2)
+
+x = Foo(bang)